Define PAR_ structure fields/SubAnim op bug fix
Newly defined particle structure values: killSphere, windMultiplier, lodReduction, lodCutoff. These values were added to the appropriate particle panels.

Some panels were updated to be a bit denser, and where applicable, the code was updated to not check list index in the draw function since that is now done in the poll function.

Updated the sub animation "Select FCurve" operation to work with the latest Blender API and cleaned up the code a bit.