Major UI update/Semi functional normal sign support
All M3 element lists now have the option to reorder items.
Most M3 element lists now have the option to duplicate items.
Added sub animations to copied data in animation duplications.

Added a new M3 panel in the Mesh property window "M3 Reverse Sign Group"
This feature is not very well implemented. If the mesh is edited in any significant way the normal sign data will become corrupt and need to be cleared or redone. As such, importing of this data is currently withheld so that people unfamiliar with the feature will not be negatively impacted by it. Improvements should be sought out.